Add explanation of fitness selection; add ordering
An individual now represents a family of datasets and its fitness is
determined by the fitness of a set of seeded samples from that family.
How to reduce this set of fitness scores down to one is up to the user
and can be done by passing such a function as the `selection` parameter
in a run of the algorithm. By default, the mean of the fitnesses is
taken to be the fitness score of that individual, but realistically, any
function that acts on a list of scores would suffice.

As well as this, the ability to choose how populations are ordered has
been added. Since some fitness functions are to be minimised, a user can
pass a boolean variable to `maximise` indicating whether fitnesses should
be maximised or not.