Filled in generation gap by taking on parents
Parents are now included in the next generation. This means the best
individuals are not destroyed after a generation.

We've found two follow-ons from this:

- Individuals seem to disappear anyway when the fitness function has an
element of randomness in it. @drvinceknight has suggested using some
smoothing here (taking the mean over a number of fitness scores). Should
be simple enough to implement.
- As individuals are being carried over into the next generation,
calculating their fitness again is a waste of time. So, caching would be
helpful in decreasing run-time. The issue with this is that
`pandas.DataFrame` objects are mutable, and so are not hashable which is
a requirement of caching functions. Maybe Dask will be the answer here?