Add test to check for reproducibility.
This tests passes but I believe this is only because the Python kernel
is the same in every run. When running tests using the circle example
from the paper (https://github.com/daffidwilde/edo-paper), different
histories are generated when restarting the kernel.

Digging around, it appears this is to do with hashing and the unordered
nature of `set()` (and potentially `dict` although this should not be
the case with Python 3.7+).

In addition to this, I'm not sure how comfortable I am about just
setting the RNG globally with `np.random.seed(global)` and forgetting
it. An attractive solution to that is to create a list of local RNGs
for the individuals in the population that they use for sampling, and
mutating. Perhaps extra RNGs will be needed for fitness calculation and
the crossover process?

If this solution is to be implemented, then classes should be built for
the global algorithm runner and each individual. It may even be the case
that the family classes need looking at.