Implement new saving strategy.
Families provide their subtypes with a `to_dict` method that records the
name of the subtype's distribution, its subtype ID number and all of its
parameters as a dictionary.

Then when an individual is saved to file, the metadata is serialized as
JSON with these dictionaries. Simultaneously, each family is cached in
the same directory tree.