Saving histories (individuals and their metadata) (#126)
* Remove `to_dict` method from `Distribution`.

* Fix the circle example test.

* Combine `Individual` to and from file tests.

* Implement new saving strategy.

Families provide their subtypes with a `to_dict` method that records the
name of the subtype's distribution, its subtype ID number and all of its
parameters as a dictionary.

Then when an individual is saved to file, the metadata is serialized as
JSON with these dictionaries. Simultaneously, each family is cached in
the same directory tree.

* Clean up optimiser with new strategy.

The optimiser can now pull usable individuals back in.

* Add a dev environment file.

* Add scipy to environment.yml

* Generalise paths for saving/loading individuals.

* [temporary] Remove parallel writing capabilities.

Because of how individuals and families are being stored, there were
some discrepancies when writing generations to file with multiple
processes. For now, that has been removed. However, the calculation of
fitnesses may still be done in parallel which is where most of the
savings likely come from anyway.

* Format codebase with isort, black and flake8.