Pickling subtypes, individuals and *families* (#127)
* Implement Family factory handler class.

* Start adding new terminology. [everything broken]

What were families are now distributions. Distribution class is much
cleaner -- is it even necessary anymore?

* Add a repr to the Family class.

* Implement the Family class in individuals.

* Hack failing distribution test.

Since subtypes are not created directly from the distribution classes,
there is no subtype_id in an instance.

* Add subtype family attribute for easy retrieval.

* Implement families in populations.

* Implement families in crossover.

* Update tests for mutation and selection.

* Implement family subtypes in shrink operator.

* Update fitness tests.

* Add Family ability to record all subtypes.

* Begin implementing families in optimiser.

This needs more attention around how individuals are saved.

* Saving histories (individuals and their metadata) (#126)

* Remove `to_dict` method from `Distribution`.

* Fix the circle example test.

* Combine `Individual` to and from file tests.

* Implement new saving strategy.

Families provide their subtypes with a `to_dict` method that records the
name of the subtype's distribution, its subtype ID number and all of its
parameters as a dictionary.

Then when an individual is saved to file, the metadata is serialized as
JSON with these dictionaries. Simultaneously, each family is cached in
the same directory tree.

* Clean up optimiser with new strategy.

The optimiser can now pull usable individuals back in.

* Add a dev environment file.

* Add scipy to environment.yml

* Generalise paths for saving/loading individuals.

* [temporary] Remove parallel writing capabilities.

Because of how individuals and families are being stored, there were
some discrepancies when writing generations to file with multiple
processes. For now, that has been removed. However, the calculation of
fitnesses may still be done in parallel which is where most of the
savings likely come from anyway.

* Format codebase with isort, black and flake8.